+++ lib/cisco_node_utils/interface.rb
@@ -18,7 +18,6 @@ require_relative 'cisco_cmn_utils'
 require_relative 'node_util'
 require_relative 'pim'
 require_relative 'vrf'
-require_relative 'vni'
 require_relative 'overlay_global'
 
 # Add some interface-specific constants to the Cisco namespace